Fix a bug in game of life demo

+++ demo-game-of-life.py
@@ -26,7 +26,7 @@ import pyglet.gl as gl
 import glumpy
 
 n = 512
-Z = numpy.random.randint(0,2,(n,n)).astype(numpy.uint8)
+Z = numpy.random.randint(0,2,(n,n)).astype(numpy.float32)
 window = pyglet.window.Window(512, 512, vsync=0)
 viewport = [0,0,1]
 Zi = glumpy.Image(Z, interpolation='nearest', cmap=glumpy.colormap.Grey,
